Update README.md
Correcting bug where () were used in the README.md file.  The correct sigils are the {} for exposing the environment variable.  In the future I could swap back to $(hostname -f) which is what was used originally, but since many servers don't have their domain set correctly, this might be better left as is.